+++ wallboxplatform.js
@@ -25,7 +25,7 @@ class wallboxPlatform {
     this.email=config.email
     this.password=config.password
     this.token
-		this.retryWait=config.retryWait || 30 //sec
+		this.retryWait=config.retryWait || 60 //sec
 		this.refreshRate=config.refreshRate || 24 //hour
 		this.liveTimeout=config.liveRefreshTimeout || 2 //min
 		this.liveRefresh=config.liveRefreshRate || 20 //sec
@@ -78,9 +78,7 @@ class wallboxPlatform {
 			this.log.debug('Found Token %s',signin.data.data.attributes.token)
 			this.id=signin.data.data.attributes.user_id
 			this.token=signin.data.data.attributes.token
-			//this.setTokenRefresh(signin.data.data.attributes.ttl-Date.now())
-			//this.log.warn(new Date(signin.data.data.attributes.ttl).toLocaleString())
-			//this.log.warn(new Date(Date.now()).toLocaleString())
+			this.setTokenRefresh(signin.data.data.attributes.ttl)
 
 			//get get user id
 			let userId=await this.wallboxapi.getId(this.token,this.id).catch(err=>{this.log.error('Failed to get userId for build', err)})
@@ -173,27 +171,30 @@ class wallboxPlatform {
 		}
 	}
 
-	setTokenRefresh(ttl){
-			if(ttl>Date.now()){
-				setInterval(async()=>{
-					try{
-						let signin=await this.wallboxapi.signin(this.email,this.password).catch(err=>{this.log.error('Failed to refresh token', err)})
-						this.log.debug('refreshed token %s',signin.data.data.attributes.token)
-						this.token=signin.data.data.attributes.token
-						this.log.info('Token has been refreshed')
-					}
-          catch(err){this.log.error('Failed to refresh token', err)}
-				}, ttl-Date.now())
-      }
-			else{
-				this.log.warn('Unable to set refresh token interval')
-			}
-		}
+  setTokenRefresh(ttl){
+    if(ttl>Date.now()){
+      ttl=ttl-Date.now()
+      setInterval(async()=>{
+        try{
+          let signin=await this.wallboxapi.signin(this.email,this.password).catch(err=>{this.log.error('Failed to refresh token', err)})
+          this.log.debug('refreshed token %s',signin.data.data.attributes.token)
+          this.id=signin.data.data.attributes.user_id
+          this.token=signin.data.data.attributes.token
+          this.setTokenRefresh(signin.data.data.attributes.ttl)
+          this.log.info('Token has been refreshed')
+        }
+        catch(err){this.log.error('Failed to refresh token', err)}
+      }, ttl)
+    }
+    else{
+      this.log.warn('Unable to set refresh token interval')
+    }
+  }
 
 	setChargerRefresh(device){
 		// Refresh charger status
 			setInterval(async()=>{
-				this.log('API polling update calls for this polling period %s',this.apiCount)
+				this.log('API calls for this polling period %s',this.apiCount)
 				this.apiCount=0
 				this.getStatus(device.id)
 			}, this.refreshRate*60*60*1000)
@@ -202,8 +203,6 @@ class wallboxPlatform {
 	async startLiveUpdate(device){
 		clearInterval(this.lastInterval)
 		//get new token
-		let signin=await this.wallboxapi.signin(this.email,this.password).catch(err=>{this.log.error('Failed to refresh token', err)})
-		this.token=signin.data.data.attributes.token
 		let startTime = new Date().getTime() //live refresh
 		if(!this.liveUpdate){this.log.debug("live update started")}
 		this.liveUpdate=true
@@ -215,9 +214,8 @@ class wallboxPlatform {
 						this.log.debug("live update stopped")
 						return
 					}
-					this.apiCount=this.apiCount+1
-					this.log.debug('API get count %s',this.apiCount)
 				this.getStatus(device.id)
+				this.log.debug('API call count %s',this.apiCount)
 			}, this.liveRefresh*1000)
 		this.lastInterval=interval
 	}